+++ Eagle-Editor/assets/shaders/screen_space_reflections/ffx_denoiser_reflections_resolve_temporal.h
@@ -31,13 +31,13 @@ shared uint g_ffx_dnsr_shared_1[16][16];
 
 struct FFX_DNSR_Reflections_NeighborhoodSample
 {
-    f16vec3 radiance;
+    vec3 radiance;
 };
 
 FFX_DNSR_Reflections_NeighborhoodSample FFX_DNSR_Reflections_LoadFromGroupSharedMemory(ivec2 idx)
 {
-    uvec2   packed_radiance   = uvec2(g_ffx_dnsr_shared_0[idx.y][idx.x], g_ffx_dnsr_shared_1[idx.y][idx.x]);
-    f16vec3 unpacked_radiance = FFX_DNSR_Reflections_UnpackFloat16_4(packed_radiance).xyz;
+    uvec2  packed_radiance = uvec2(g_ffx_dnsr_shared_0[idx.y][idx.x], g_ffx_dnsr_shared_1[idx.y][idx.x]);
+    vec3 unpacked_radiance = FFX_DNSR_Reflections_UnpackFloat16_4(packed_radiance).xyz;
 
     FFX_DNSR_Reflections_NeighborhoodSample smpl;
     smpl.radiance = unpacked_radiance;
@@ -46,23 +46,23 @@ FFX_DNSR_Reflections_NeighborhoodSample FFX_DNSR_Reflections_LoadFromGroupShared
 
 struct FFX_DNSR_Reflections_Moments
 {
-    f16vec3 mean;
-    f16vec3 variance;
+    vec3 mean;
+    vec3 variance;
 };
 
 FFX_DNSR_Reflections_Moments FFX_DNSR_Reflections_EstimateLocalNeighborhoodInGroup(ivec2 group_thread_id)
 {
     FFX_DNSR_Reflections_Moments estimate;
-    estimate.mean                = f16vec3(0);
-    estimate.variance            = f16vec3(0);
-    float16_t accumulated_weight = float16_t(0);
+    estimate.mean                = vec3(0);
+    estimate.variance            = vec3(0);
+    float accumulated_weight = 0;
     for (int j = -FFX_DNSR_REFLECTIONS_LOCAL_NEIGHBORHOOD_RADIUS; j <= FFX_DNSR_REFLECTIONS_LOCAL_NEIGHBORHOOD_RADIUS; ++j)
     {
         for (int i = -FFX_DNSR_REFLECTIONS_LOCAL_NEIGHBORHOOD_RADIUS; i <= FFX_DNSR_REFLECTIONS_LOCAL_NEIGHBORHOOD_RADIUS; ++i)
         {
             ivec2     new_idx  = group_thread_id + ivec2(i, j);
-            f16vec3   radiance = FFX_DNSR_Reflections_LoadFromGroupSharedMemory(new_idx).radiance;
-            float16_t weight   = FFX_DNSR_Reflections_LocalNeighborhoodKernelWeight(float16_t(i)) * FFX_DNSR_Reflections_LocalNeighborhoodKernelWeight(float16_t(j));
+            vec3   radiance    = FFX_DNSR_Reflections_LoadFromGroupSharedMemory(new_idx).radiance;
+            float weight       = FFX_DNSR_Reflections_LocalNeighborhoodKernelWeight(float(i)) * FFX_DNSR_Reflections_LocalNeighborhoodKernelWeight(float(j));
             accumulated_weight += weight;
             estimate.mean      += radiance * weight;
             estimate.variance  += radiance * radiance * weight;
@@ -75,13 +75,13 @@ FFX_DNSR_Reflections_Moments FFX_DNSR_Reflections_EstimateLocalNeighborhoodInGro
     return estimate;
 }
 
-void FFX_DNSR_Reflections_StoreInGroupSharedMemory(ivec2 group_thread_id, f16vec3 radiance)
+void FFX_DNSR_Reflections_StoreInGroupSharedMemory(ivec2 group_thread_id, vec3 radiance)
 {
     g_ffx_dnsr_shared_0[group_thread_id.y][group_thread_id.x] = FFX_DNSR_Reflections_PackFloat16(radiance.xy);
     g_ffx_dnsr_shared_1[group_thread_id.y][group_thread_id.x] = FFX_DNSR_Reflections_PackFloat16(radiance.zz);
 }
 
-void FFX_DNSR_Reflections_LoadNeighborhood(ivec2 pixel_coordinate, out f16vec3 radiance)
+void FFX_DNSR_Reflections_LoadNeighborhood(ivec2 pixel_coordinate, out vec3 radiance)
 {
     radiance = FFX_DNSR_Reflections_LoadRadiance(pixel_coordinate);
 }
@@ -92,7 +92,7 @@ void FFX_DNSR_Reflections_InitializeGroupSharedMemory(ivec2 dispatch_thread_id, 
     ivec2 offset[4] = {ivec2(0, 0), ivec2(8, 0), ivec2(0, 8), ivec2(8, 8)};
 
     // Intermediate storage registers to cache the result of all loads
-    f16vec3 radiance[4];
+    vec3 radiance[4];
 
     // Start in the upper left corner of the 16x16 region.
     dispatch_thread_id -= 4;
@@ -119,32 +119,32 @@ void FFX_DNSR_Reflections_ResolveTemporal(ivec2 dispatch_thread_id, ivec2 group_
     group_thread_id += 4; // Center threads in groupshared memory
 
     FFX_DNSR_Reflections_NeighborhoodSample center       = FFX_DNSR_Reflections_LoadFromGroupSharedMemory(group_thread_id);
-    f16vec3                                 new_signal   = center.radiance;
-    float16_t                               roughness    = FFX_DNSR_Reflections_LoadRoughness(dispatch_thread_id);
-    float16_t                               new_variance = FFX_DNSR_Reflections_LoadVariance(dispatch_thread_id);
+    vec3                                    new_signal   = center.radiance;
+    float                                   roughness    = FFX_DNSR_Reflections_LoadRoughness(dispatch_thread_id);
+    float                                   new_variance = FFX_DNSR_Reflections_LoadVariance(dispatch_thread_id);
     
     if (FFX_DNSR_Reflections_IsGlossyReflection(float(roughness)))
     {
-        float16_t  num_samples  = FFX_DNSR_Reflections_LoadNumSamples(dispatch_thread_id);
-        vec2       uv8          = (vec2(dispatch_thread_id.xy) + (0.5).xx) / FFX_DNSR_Reflections_RoundUp8(screen_size);
-        f16vec3    avg_radiance = FFX_DNSR_Reflections_SampleAverageRadiance(uv8);
+        float  num_samples  = FFX_DNSR_Reflections_LoadNumSamples(dispatch_thread_id);
+        vec2   uv8          = (vec2(dispatch_thread_id.xy) + (0.5).xx) / FFX_DNSR_Reflections_RoundUp8(screen_size);
+        vec3   avg_radiance = FFX_DNSR_Reflections_SampleAverageRadiance(uv8);
 
-        f16vec3                      old_signal         = FFX_DNSR_Reflections_LoadRadianceReprojected(dispatch_thread_id);
+        vec3                      old_signal         = FFX_DNSR_Reflections_LoadRadianceReprojected(dispatch_thread_id);
         FFX_DNSR_Reflections_Moments local_neighborhood = FFX_DNSR_Reflections_EstimateLocalNeighborhoodInGroup(group_thread_id);
         // Clip history based on the curren local statistics
-        f16vec3                  color_std          = (sqrt(local_neighborhood.variance.xyz) + length(local_neighborhood.mean.xyz - avg_radiance)) * float16_t(history_clip_weight * 1.4);
-                                 local_neighborhood.mean.xyz = mix(local_neighborhood.mean.xyz, avg_radiance, f16vec3(0.2));
-        f16vec3                  radiance_min       = local_neighborhood.mean.xyz - color_std;
-        f16vec3                  radiance_max       = local_neighborhood.mean.xyz + color_std;
-        f16vec3                  clipped_old_signal = FFX_DNSR_Reflections_ClipAABB(radiance_min, radiance_max, old_signal.xyz);
-        float16_t                accumulation_speed = float16_t(1.0) / max(num_samples, float16_t(1.0));
-        float16_t                weight             = (float16_t(1.0) - accumulation_speed);
+        vec3         color_std          = (sqrt(local_neighborhood.variance.xyz) + length(local_neighborhood.mean.xyz - avg_radiance)) * (history_clip_weight * 1.4);
+                     local_neighborhood.mean.xyz = mix(local_neighborhood.mean.xyz, avg_radiance, vec3(0.2));
+        vec3         radiance_min       = local_neighborhood.mean.xyz - color_std;
+        vec3         radiance_max       = local_neighborhood.mean.xyz + color_std;
+        vec3         clipped_old_signal = FFX_DNSR_Reflections_ClipAABB(radiance_min, radiance_max, old_signal.xyz);
+        float        accumulation_speed = 1.0 / max(num_samples, 1.0);
+        float        weight             = (1.0 - accumulation_speed);
         // Blend with average for small sample count
-        new_signal.xyz                                  = mix(new_signal.xyz, avg_radiance, float16_t(1.0) / max(num_samples + float16_t(1.0f), float16_t(1.0)));
+        new_signal.xyz                                  = mix(new_signal.xyz, avg_radiance, 1.0 / max(num_samples + 1.0f, 1.0));
         // Clip outliers
         {
-            f16vec3 radiance_min = avg_radiance.xyz - color_std * float16_t(1.0);
-            f16vec3 radiance_max = avg_radiance.xyz + color_std * float16_t(1.0);
+            vec3 radiance_min = avg_radiance.xyz - color_std * 1.0;
+            vec3 radiance_max = avg_radiance.xyz + color_std * 1.0;
             new_signal.xyz       = FFX_DNSR_Reflections_ClipAABB(radiance_min, radiance_max, new_signal.xyz);
         }
         // Blend with history
@@ -152,8 +152,8 @@ void FFX_DNSR_Reflections_ResolveTemporal(ivec2 dispatch_thread_id, ivec2 group_
         new_variance = mix(FFX_DNSR_Reflections_ComputeTemporalVariance(new_signal.xyz, clipped_old_signal.xyz), new_variance, weight);
         if (any(isinf(vec3(new_signal))) || any(isnan(vec3(new_signal))) || any(isinf(vec3(new_variance))) || any(isnan(vec3(new_variance))))
         {
-            new_signal   = f16vec3(0.0);
-            new_variance = float16_t(0.0);
+            new_signal   = vec3(0.0);
+            new_variance = 0.0;
         }
 
     }
